+++ src/hipay_enterprise/controllers/front/redirect.php
@@ -10,6 +10,7 @@
  * @copyright 2017 HiPay
  * @license   https://github.com/hipay/hipay-enterprise-sdk-prestashop/blob/master/LICENSE.md
  */
+require_once dirname(__FILE__).'/../../lib/vendor/autoload.php';
 require_once dirname(__FILE__).'/../../classes/apiHandler/ApiHandler.php';
 require_once dirname(__FILE__).'/../../classes/helper/HipayHelper.php';
 require_once dirname(__FILE__).'/../../classes/helper/HipayCCToken.php';
@@ -199,9 +200,19 @@ class Hipay_enterpriseRedirectModuleFrontController extends ModuleFrontControlle
                     $this->assignTemplate();
                     $path = 'payment/ps16/paymentForm-'.$uxMode.'-16.tpl';
                 } else {
+                    // Impossible case but necessary
                     $this->assignTemplate();
+                    $this->context->smarty->assign(
+                        [
+                            'HiPay_action' => $this->context->link->getModuleLink(
+                                $this->module->name, 'redirect', [], true
+                            ),
+                            'HiPay_languageIsoCode' => $this->context->language->iso_code,
+                        ]
+                    );
                     $path = _PS_VERSION_ >= '1.7' ?
-                        'payment/ps17/paymentForm-'.$uxMode.'-17.tpl'
+                        'module:'.$this->module->name.
+                        '/views/templates/front/payment/ps17/paymentForm-'.$uxMode.'-17.tpl'
                         : 'payment/ps16/paymentForm-'.$uxMode.'-16.tpl';
                 }
                 break;